% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-elapsed-time.R
\name{check_elapsed_time}
\alias{check_elapsed_time}
\title{Printing for timing processes}
\usage{
check_elapsed_time(last_time, start_time)
}
\arguments{
\item{last_time}{**optional** object of class \code{POSIXct}, indicating the most recent check. Usually the results of this function}

\item{start_time}{**optional** object of class \code{POSIXct}, indicating the beginning of the run.}
}
\value{
results of \code{\link{Sys.time()}}
}
\description{
Creates a set of printed results for displaying the time it takes to run a long process
Rounds to the nearest hundreth of a second, so not suitible for very short processes
Note that this \code{print}s. Do not use unless you are expecting that.
}
\examples{
ptm <- Sys.time()
Sys.sleep(4)
checkpoint <- check_elapsed_time(start_time = ptm)
Sys.sleep(8)
checkpoint <- check_elapsed_time(last_time = checkpoint, start_time = ptm)
}
