% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{count_seasons}
\alias{count_seasons}
\title{Count the number of seasons between two termkeys}
\usage{
count_seasons(termkey1, termkey2, neg = TRUE, ignore_summers = FALSE)
}
\arguments{
\item{termkey1}{A Termkey (will be validated). By convention, should be earlier}

\item{termkey2}{A Termkey (will be validated). By convention, should be later}

\item{neg}{Boolean: return a negative number of seasons if termkey 1 is later than termkey2, if FALSE, provide an absolute difference}

\item{ignore_summers}{Boolean: if TRUE, count fall and spring semesters only}
}
\description{
Fall, spring, summer.
If ignore_summers = TRUE, but one of the termkeys is a summer, it will make the summer term collapse into the previous spring. 
This gives an accurate count of the number of fall and spring terms in between two termkeys.
}
