% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{find_sample}
\alias{find_sample}
\title{Find a sample common to two sets}
\usage{
find_sample(new_vec, old_sample = list(vec = NULL, sample = NULL),
  size_pct = 0.002)
}
\description{
Find a sample common to two sets
}
