% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpg.R
\name{gpg_rds_decrypt}
\alias{gpg_rds_decrypt}
\title{Load encrypted data}
\usage{
gpg_rds_decrypt(encrypted_file)
}
\arguments{
\item{encrypted_file}{Path to an encrypted file made using this package}
}
\value{
Original object stored in file
}
\description{
Load encrypted data
}
\examples{
path <- gpg_rds_encrypt(mtcars, recipient = "AAxthelm@che.in.gov")
mtc2 <- gpg_rds_decrypt(path)
stopifnot(identical(mtc2, mtcars))
}
