% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpg.R
\name{gpg_rds_encrypt}
\alias{gpg_rds_encrypt}
\title{This function does not accept vectorized inputs. It will only deal with whole objects}
\usage{
gpg_rds_encrypt(x, recipient, ..., path = "data", overwrite = TRUE,
  obj_name = NULL)
}
\description{
This function does not accept vectorized inputs. It will only deal with whole objects
}
