% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted-data.R
\name{make_tiny_frame}
\alias{make_tiny_frame}
\title{Create a subset of data
TODO: Expand this documentation}
\usage{
make_tiny_frame(frame_df, keycolumn, randomvec, sample_n = 5)
}
\arguments{
\item{frame_df}{a dataframe or similar (tbl)}

\item{keycolumn}{character: name of column by which to sample. If missing, defaults to row names}

\item{randomvec}{a vector a values in keycolumn to return}

\item{sample_n}{number of samples to return, if randomvec is missing}
}
\value{
a subsetted object of same type as frame_df
}
\description{
Create a subset of data
TODO: Expand this documentation
}
\examples{
make_tiny_frame(mtcars)
make_tiny_frame(ChickWeight, keycolumn = "Chick")
}
