% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted-data.R
\name{read_gpg_frame_and_tiny}
\alias{read_gpg_frame_and_tiny}
\title{Load an encrypted RDS
TODO: Expand this Documentation}
\usage{
read_gpg_frame_and_tiny(encrypted_file, use_tiny = TRUE, failsafe = TRUE)
}
\arguments{
\item{encrypted_file}{foobar}

\item{use_tiny}{foobar}

\item{failsafe}{foobar}
}
\value{
foobar
}
\description{
Load an encrypted RDS
TODO: Expand this Documentation
}
\examples{
TRUE
}
