% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{read_SQL_query_from_file}
\alias{read_SQL_query_from_file}
\title{Turn an SQL file into a query the DBI can understand}
\usage{
read_SQL_query_from_file(file, printquery = FALSE)
}
\description{
Turn an SQL file into a query the DBI can understand
}
