% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_name}
\alias{sample_name}
\title{Create a name suitible for feeding into vault_save}
\usage{
sample_name(name, use_sample = config$use_sample,
  sample_pct = config$sample_pct)
}
\description{
Create a name suitible for feeding into vault_save
}
