% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted-data.R
\name{save_frame_and_tiny}
\alias{save_frame_and_tiny}
\title{Save an encrypted RDS
TODO: Expand this Documentation}
\usage{
save_frame_and_tiny(frame_df, keycolumn, randomvec, path = "data",
  make_tiny_args = list())
}
\arguments{
\item{frame_df}{data.frame, or similar}

\item{keycolumn}{column of frame_df to select from}

\item{randomvec}{vector of values that are in keycolumn}
}
\value{
foobar
}
\description{
Save an encrypted RDS
TODO: Expand this Documentation
}
\examples{
TRUE
}
