% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted-data.R
\name{save_gpg_frame_and_tiny}
\alias{save_gpg_frame_and_tiny}
\title{Save an encrypted RDS
TODO: Expand this Documentation}
\usage{
save_gpg_frame_and_tiny(frame_df, recipient, ..., make_tiny_args = list(...),
  gpg_rds_encrypt_args = list(...))
}
\arguments{
\item{frame_df}{foobar}

\item{recipient}{foobar}

\item{...}{...}

\item{make_tiny_args}{foobar}

\item{gpg_rds_encrypt_args}{foobar}
}
\value{
foobar
}
\description{
Save an encrypted RDS
TODO: Expand this Documentation
}
\examples{
TRUE
}
