% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term.R
\name{split_termkey}
\alias{split_termkey}
\title{Split Termkey into important information}
\usage{
split_termkey(termkey)
}
\arguments{
\item{termkey}{}
}
\value{
List:
academic_year,
term_season,
term_season_number,
termcode,
term_name,
fiscal_year
}
\description{
Split Termkey into important information
}
\examples{
split_termkey(20083)
split_termkey("20291")
}
