% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted-data.R
\name{vault_frame_and_tiny}
\alias{vault_frame_and_tiny}
\title{Put a frame and its subset into the `secret` vault
TODO: Expand this documentation}
\usage{
vault_frame_and_tiny(frame_df, keycolumn, randomvec, vault, users)
}
\arguments{
\item{frame_df}{a dataframe or similar (tbl)}

\item{keycolumn}{character: name of column by which to sample. If missing, defaults to row names}

\item{randomvec}{a vector a values in keycolumn to return}

\item{vault}{the vault to which to save}
}
\value{
a subsetted object of same type as frame_df
}
\description{
Put a frame and its subset into the `secret` vault
TODO: Expand this documentation
}
