% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_elapsed_time.R
\name{check_elapsed_time}
\alias{check_elapsed_time}
\title{Messages for timing processes}
\usage{
check_elapsed_time(last_time, start_time)
}
\arguments{
\item{last_time}{**optional** object of class \code{proc_time}, indicating the most recent check. Usually the results of this function}

\item{start_time}{**optional** object of class \code{proc_time}, indicating the beginning of the run.}
}
\value{
results of \code{\link{proc.time}}
}
\description{
Creates a set of messages for displaying the time it takes to run a long process
Rounds to the nearest hundreth of a second, so not suitible for very short processes
}
