% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpg.R
\name{gpg_rds_encrypt}
\alias{gpg_rds_encrypt}
\title{Save an R object as an encrypted file}
\usage{
gpg_rds_encrypt(x, recipient, ..., path = "data", overwrite = TRUE,
  obj_name = NULL)
}
\arguments{
\item{x}{any R object}

\item{recipient}{the GPG recipient (passes args to \code{-r})}

\item{...}{...}

\item{path}{Path to which file should be written}

\item{overwrite}{If TRUE, will overwrite existing file with same name}

\item{obj_name}{if not NULL, will use string as filename}
}
\value{
Path to encrypted file
}
\description{
Save an R object as an encrypted file
}
\examples{
\dontrun{ 
path <- gpg_rds_encrypt(mtcars, recipient = "AAxthelm@che.in.gov")
mtc2 <- gpg_rds_decrypt(path)
stopifnot(identical(mtc2, mtcars))
}
}
