% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetted-data.R
\name{make_tiny_frame}
\alias{make_tiny_frame}
\title{Create a subset of data
TODO: Expand this documentation}
\usage{
make_tiny_frame(frame_df, ..., keycolumn = "CurrentCSN", randomvec = NULL,
  sample_size = 0.01)
}
\arguments{
\item{frame_df}{foobar}

\item{...}{...}

\item{keycolumn}{foobar}

\item{randomvec}{foobar}

\item{sample_size}{foobar}
}
\value{
foobar
}
\description{
Create a subset of data
TODO: Expand this documentation
}
\examples{
TRUE
}
